Decoupled room acoustic code from Juce.
Introduced more issues with locks that needs to be cleaned up.

+++ SignalProcessing.cpp
@@ -7,9 +7,9 @@
 
 using namespace NicDSP;
 
-void Butterworth1Pole::Initialize(double cutoff_frequency, double sample_rate)
+void Butterworth1Pole::Initialize(float cutoff_frequency, float sample_rate)
 {
-    const float blt_freq_warping = 1 / tanf(M_PI * cutoff_frequency / sample_rate);
+    const float blt_freq_warping = 1 / tanf((float)M_PI * cutoff_frequency / sample_rate);
 
     const float a0 = 1 + blt_freq_warping;
     a1 = (1 - blt_freq_warping) / a0;
